combine day and dinner slot comps
